Abstract out validateURI method()
