+++ core/client/hdfs/src/main/java/alluxio/hadoop/FileSystem.java
@@ -14,8 +14,15 @@ package alluxio.hadoop;
 import alluxio.Constants;
 import alluxio.conf.PropertyKey;
 import alluxio.annotation.PublicApi;
+import alluxio.exception.PreconditionMessage;
+import alluxio.uri.Authority;
+import alluxio.uri.UnknownAuthority;
+
+import com.google.common.base.Preconditions;
 
 import javax.annotation.concurrent.NotThreadSafe;
+import java.io.IOException;
+import java.net.URI;
 
 /**
  * An Alluxio client API compatible with Apache Hadoop {@link org.apache.hadoop.fs.FileSystem}
@@ -51,5 +58,17 @@ public final class FileSystem extends AbstractFileSystem {
   @Override
   protected boolean isZookeeperMode() {
     return mFileSystem.getConf().getBoolean(PropertyKey.ZOOKEEPER_ENABLED);
+  }
+
+  @Override
+  protected void validateFsUri(URI fsUri) throws IOException, IllegalArgumentException {
+    Preconditions.checkArgument(fsUri.getScheme().equals(getScheme()),
+        PreconditionMessage.URI_SCHEME_MISMATCH.toString(), fsUri.getScheme(), getScheme());
+
+    Authority auth = Authority.fromString(fsUri.getAuthority());
+    if (auth instanceof UnknownAuthority) {
+      throw new IOException(String.format("Authority \"%s\" is unknown. The client can not be "
+          + "configured with the authority from %s", auth, fsUri));
+    }
   }
 }
